backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
pandas==2.1.3
numpy==1.24.4
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.25.2
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
